+++ apps/kitchen-sink/src/ensemble/scripts/test.js
@@ -1,3 +1,3 @@
 const sayHello = () => {
-  window.alert("hello world!")
-}
+  window.alert("hello world!");
+};