+++ packages/framework/src/evaluate/evaluate.ts
@@ -6,7 +6,7 @@ import { sanitizeJs, debug } from "../shared";
 export const buildEvaluateFn = (
   screen: Partial<ScreenContextDefinition>,
   js?: string,
-  context?: Record<string, unknown>,
+  context?: { [key: string]: unknown },
 ): (() => unknown) => {
   const widgets: [string, InvokableMethods | undefined][] = Object.entries(
     screen.widgets ?? {},
@@ -20,14 +20,16 @@ export const buildEvaluateFn = (
     ...widgets,
     ...Object.entries(screen.inputs ?? {}),
     ...Object.entries(screen.data ?? {}),
+    ...Object.entries(screen),
     ...Object.entries(context ?? {}),
   ]);
   const globalBlock = screen.model?.global;
+  const importedScriptBlock = screen.model?.importedScripts;
 
   // eslint-disable-next-line @typescript-eslint/no-implied-eval, no-new-func
   const jsFunc = new Function(
     ...Object.keys(invokableObj),
-    addGlobalBlock(formatJs(js), globalBlock),
+    addScriptBlock(formatJs(js), globalBlock, importedScriptBlock),
   );
 
   // eslint-disable-next-line @typescript-eslint/no-unsafe-return
@@ -63,8 +65,23 @@ const formatJs = (js?: string): string => {
   return `return ${sanitizedJs}`;
 };
 
-const addGlobalBlock = (js: string, globalBlock?: string): string =>
-  globalBlock ? `${globalBlock}\n\n${js}` : js;
+const addScriptBlock = (
+  js: string,
+  globalBlock?: string,
+  importedScriptBlock?: string,
+): string => {
+  let jsString = ``;
+
+  if (importedScriptBlock) {
+    jsString += `${importedScriptBlock}\n\n`;
+  }
+
+  if (globalBlock) {
+    jsString += `${globalBlock}\n\n`;
+  }
+
+  return (jsString += `${js}`);
+};
 
 /**
  * @deprecated Consider using useEvaluate or createBinding which will
@@ -78,7 +95,7 @@ const addGlobalBlock = (js: string, globalBlock?: string): string =>
 export const evaluate = <T = unknown>(
   screen: Partial<ScreenContextDefinition>,
   js?: string,
-  context?: Record<string, unknown>,
+  context?: { [key: string]: unknown },
 ): T => {
   try {
     return buildEvaluateFn(screen, js, context)() as T;