+++ packages/framework/src/hooks/useRegisterBindings.ts
@@ -1,12 +1,12 @@
 import type { RefCallback } from "react";
-import { useEffect } from "react";
-import { get, isEmpty, isString, keys, merge } from "lodash-es";
+import { useEffect, useMemo } from "react";
+import { get, isEmpty, isString, keys, debounce } from "lodash-es";
 import isEqual from "react-fast-compare";
 import type { InvokableMethods } from "../state";
 import { useWidgetId } from "./useWidgetId";
 import { useHtmlPassThrough } from "./useHtmlPassThrough";
 import { useWidgetState } from "./useWidgetState";
-import { useStyleNames } from "./useStyleNames";
+import { useStyles } from "./useStyles";
 import { useEvaluate } from "./useEvaluate";
 
 export interface RegisterBindingsResult<T> {
@@ -21,6 +21,7 @@ export const useRegisterBindings = <T extends { [key: string]: unknown }>(
   methods?: InvokableMethods,
   options?: {
     forceState?: boolean;
+    debounceMs?: number;
   },
 ): RegisterBindingsResult<T> => {
   const testId = get(values, ["testId"]);
@@ -32,20 +33,19 @@ export const useRegisterBindings = <T extends { [key: string]: unknown }>(
 
   const [widgetState, setWidgetState] = useWidgetState<T>(resolvedWidgetId);
 
-  const namedStyles = get(values, ["styles", "names"]) as unknown;
-  const classStyles = get(values, ["styles", "className"]) as unknown;
+  const styles = useStyles(values);
 
-  const styleProperties = useStyleNames(
-    isString(namedStyles) ? namedStyles : "",
-    isString(classStyles) ? classStyles : "",
+  const newValues = useEvaluate(
+    { ...values, ...(isEmpty(styles) ? {} : { styles }) },
+    {
+      debugId: resolvedWidgetId,
+    },
   );
 
-  const styles = merge({}, styleProperties, values.styles);
-  if (!isEmpty(styles)) {
-    merge(values, { styles });
-  }
-
-  const newValues = useEvaluate(values, { debugId: resolvedWidgetId });
+  const debounceSetState = useMemo(
+    () => debounce(setWidgetState, options?.debounceMs ?? 0, { leading: true }),
+    [options?.debounceMs, setWidgetState],
+  );
 
   useEffect(() => {
     // Improves performance greatly: o need to store state in global if there's no explicit ID to reference it with
@@ -60,7 +60,7 @@ export const useRegisterBindings = <T extends { [key: string]: unknown }>(
       return;
     }
 
-    setWidgetState({
+    debounceSetState({
       values: newValues,
       invokable: { id: resolvedWidgetId, methods },
     });
@@ -72,6 +72,7 @@ export const useRegisterBindings = <T extends { [key: string]: unknown }>(
     widgetState?.values,
     widgetState?.invokable.methods,
     id,
+    debounceSetState,
   ]);
 
   useEffect(() => {