+++ apps/kitchen-sink/CHANGELOG.md
@@ -1,5 +1,89 @@
 # @ensembleui/react-kitchen-sink
 
+## 0.0.38
+
+### Patch Changes
+
+- 02dca79: fix initial value of ToggleButton
+- 3c5c191: Make Search widget event symmetrical with data binding
+- 38d10e1: expose context in executeActionGroup
+- Updated dependencies [02dca79]
+- Updated dependencies [3c5c191]
+- Updated dependencies [38d10e1]
+  - @ensembleui/react-runtime@0.1.47
+
+## 0.0.37
+
+### Patch Changes
+
+- 8a0b3a1: added defaultSelected Row in datagrid
+- 0730867: Added support for import scripts
+- c923f2a: fix search input infinite render issue
+- Updated dependencies [8a0b3a1]
+- Updated dependencies [0730867]
+- Updated dependencies [c923f2a]
+  - @ensembleui/react-runtime@0.1.46
+  - @ensembleui/react-framework@0.1.34
+
+## 0.0.36
+
+### Patch Changes
+
+- df66279: support for radio type in DataGrid
+- f801ba7: show empty progress bar when filledPercentage is 0
+- Updated dependencies [df66279]
+- Updated dependencies [f801ba7]
+  - @ensembleui/react-runtime@0.1.45
+
+## 0.0.35
+
+### Patch Changes
+
+- f1cf9b0: added support of expression in style and class names
+- 593537d: Added support for onPageChange action handler
+- 8072925: Add functionality to disable create new options in multiselect
+- 12eada9: add humanize method to ensemble formatter
+- 8b6a691: Add initial support for executeActionGroup
+- 3cdf0db: fix navigateExternalScreen API to accept the args from the other actions
+- 887f10a: set value of Radio as undefined instead of "undefined"
+- Updated dependencies [f1cf9b0]
+- Updated dependencies [593537d]
+- Updated dependencies [8072925]
+- Updated dependencies [12eada9]
+- Updated dependencies [8b6a691]
+- Updated dependencies [4c5cba2]
+- Updated dependencies [3cdf0db]
+- Updated dependencies [887f10a]
+- Updated dependencies [87876d7]
+  - @ensembleui/react-framework@0.1.33
+  - @ensembleui/react-runtime@0.1.44
+
+## 0.0.34
+
+### Patch Changes
+
+- 3b61b34: receive updated values from all form elements
+- Updated dependencies [3b61b34]
+  - @ensembleui/react-runtime@0.1.43
+
+## 0.0.33
+
+### Patch Changes
+
+- Updated dependencies [357d754]
+  - @ensembleui/react-framework@0.1.32
+  - @ensembleui/react-runtime@0.1.42
+
+## 0.0.32
+
+### Patch Changes
+
+- Updated dependencies [3da9f12]
+- Updated dependencies [f566152]
+- Updated dependencies [f566152]
+  - @ensembleui/react-runtime@0.1.41
+  - @ensembleui/react-framework@0.1.31
+
 ## 0.0.31
 
 ### Patch Changes